Правки сущностей [#4745 #417625]

+++ src/main/java/ru/buycar/entity/DoorCount.java
@@ -1,7 +1,6 @@
 package ru.buycar.entity;
 
 import lombok.*;
-
 import javax.persistence.*;
 
 @AllArgsConstructor
@@ -12,12 +11,12 @@ import javax.persistence.*;
 @Table(name = "door_count")
 @Entity
 public class DoorCount {
-    //todo добавить таблицу
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(name = "amount")
-    private byte amount;
+    private int amount;
 
 }
